Extract Prompt Studio extend slot helpers
