+++ docs/development/issue-14-frontend-js-maintenance.md
@@ -208,6 +208,7 @@ web/js/
     textarea.js
     wheel.js
     dom.js
+    extend_slots.js
     style.js
     tooltip.js
     widgets.js
@@ -234,13 +235,15 @@ Planned slices:
 8. `layout.js`
 9. `textarea.js`
 10. `wheel.js`
-11. `dom.js` and `style.js`
-12. `tooltip.js`
-13. `widgets.js`
-14. `legend.js`
-15. `fields.js`
-16. `node_hooks.js`
-17. entry file slim-down
+11. `dom.js`
+12. `extend_slots.js`
+13. `style.js`
+14. `tooltip.js`
+15. `widgets.js`
+16. `legend.js`
+17. `fields.js`
+18. `node_hooks.js`
+19. entry file slim-down
 
 Each PR should mostly move code and update imports. Behavior changes should be
 separate PRs unless the move reveals a confirmed bug that cannot be separated.
@@ -718,6 +721,7 @@ Use this in Issue #14 or follow-up PR tracking comments.
 [ ] textarea.js extracted
 [ ] wheel.js extracted
 [ ] dom.js extracted
+[ ] extend_slots.js extracted
 [ ] style.js extracted
 [ ] tooltip.js extracted
 [ ] widgets.js extracted
